+++ frontend/src/pages/NotFoundPage/styled.tsx
